requirements.txt
APScheduler==3.6.3
backports.entry-points-selectable==1.1.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
contextlib2==21.6.0
datedelta==1.3
distlib==0.3.3
filelock==3.0.12
Flask==2.0.1
Flask-Login==0.5.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.3
platformdirs==2.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyTelegramBotAPI==4.1.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
schema==0.7.4
six==1.16.0
snaql==0.4.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
virtualenv==20.8.1
Werkzeug==2.0.1
